moved dsc commands to own plugin
