+++ plugins/misc.py
@@ -1,46 +1,20 @@
-import random
 import discord
-
-from datetime import datetime
 from discord.ext import commands
 from conf import Config
-from botutils import utils, permChecks
-from botutils.enums import DscState
+
 from Geckarbot import BasePlugin
 
 
 class Plugin(BasePlugin, name="Funny/Misc Commands"):
-    """Funny and miscellaneous commands without other category"""
-
-    songmaster_role_id = 0
+    """Sport related commands"""
 
     def __init__(self, bot):
         self.bot = bot
         super().__init__(bot)
         bot.register(self)
 
-        global songmaster_role_id
-        songmaster_role_id = Config().ROLE_IDS.get('songmaster', 0)
-
     def default_config(self):
-        return {
-            'dsc':{
-                'rule_link': "https://docs.google.com/document/d/1xvkIPgLfFvm4CLwbCoUa8WZ1Fa-Z_ELPAtgHaSpEEbg",
-                'contestdoc_link': "https://docs.google.com/spreadsheets/d/1HH42s5DX4FbuEeJPdm8l1TK70o2_EKADNOLkhu5qRa8",
-                'host_id': None,
-                'state': DscState.NA,
-                'yt_link': None,
-                'state_end': datetime.now()
-                }
-            }
-
-    def dsc_conf(self):
-        return Config().get(self)['dsc']
-
-
-    ######
-    # Simple games
-    ######
+        return {}
 
     @commands.command(name="dice", brief="Simulates rolling dice.",
                       usage="[NumberOfSides] [NumberOfDices]")
@@ -52,111 +26,46 @@ class Plugin(BasePlugin, name="Funny/Misc Commands"):
         ]
         await ctx.send(', '.join(dice))
 
-    ######
-    # DSC
-    ######
+    @commands.command(name="kicker", help="Returns frequently used links to kicker.de")
+    async def kicker_table(self, ctx):
+        """Returns the kicker.de Bundesliga tables"""
+        embed = discord.Embed(title=Config().lang(self, 'kicker_title'))
+        embed.add_field(name="Bundesliga", value="https://www.kicker.de/1-bundesliga/tabelle")
+        embed.add_field(name="2. Bundesliga", value="https://www.kicker.de/2-bundesliga/tabelle")
+        embed.add_field(name="3. Liga", value="https://www.kicker.de/3-liga/tabelle")
+        embed.add_field(name="AT-Bundesliga", value="https://www.kicker.de/tipp3-bundesliga/tabelle\n"
+                                                     "https://www.kicker.de/meistergruppe-oesterreich-fb-1/tabelle/\n"
+                                                     "https://www.kicker.de/qualifikationsgruppe-oesterreich-fb-1/tabelle/")
+        await ctx.send(embed=embed)
 
-    @commands.group(name="dsc", help="Get and manage informations about current DSC",
-                    description="Get the informations about the current dsc or manage it. "
-                                "Command only works in music channel. "
-                                "Manage DSC informations is only permitted for songmasters.")
-    @permChecks.in_channel(Config().CHAN_IDS.get('music', 0))
-    async def dsc(self, ctx):
-        """DSC base command, return info command if no subcommand given"""
-        if ctx.invoked_subcommand is None:
-            await self.dsc_get_info(ctx)
+    @commands.command(name="ping", help="Pings the bot.")
+    async def ping(self, ctx):
+        await ctx.send("Pong!")
+        
+    @commands.command(name="mud", brief="Pings the bot.")
+    async def mud(self, ctx):
+        await ctx.send("Kip!")
+        
+    @commands.command(name="mudkip", brief="MUDKIP!")
+    async def mudkip(self, ctx):
+        await ctx.send("https://www.youtube.com/watch?v=3DkqMjfqqPc")
 
-    @dsc.command(name="rules", help="Get the link to the DSC rules")
-    async def dsc_get_rules(self, ctx):
-        """Returns the DSC rules"""
-        await ctx.send(f"<{self.dsc_conf()['rule_link']}>")
+    @commands.command(name="nico", help="Punches Nico.")
+    async def nico(self, ctx):
+        await ctx.send(Config().lang(self, 'nico_output'))
 
-    @dsc.command(name="info", help="Get informations about current DSC")
-    async def dsc_get_info(self, ctx):
-        """Returns basic infos about next/current DSC"""
-        hostNick = None
-        dateOutStr = ""
-        if not self.dsc_conf()['host_id']:
-            await ctx.send("You must set DSC host!")
-        else:
-            hostNick = discord.utils.get(ctx.guild.members, id=self.dsc_conf()['host_id']).name
+    @commands.command(name="mimimi", help="Provides an .mp3 file that plays the sound of 'mimimi'.")
+    async def mimimi(self, ctx):
+        await ctx.trigger_typing()
+        file = discord.File(f"{Config().storage_dir(self)}/mimimi.mp3")
+        await ctx.send(file=file)
 
-        if self.dsc_conf()['state'] == DscState.Registration:
-            if self.dsc_conf()['state_end'] > datetime.now():
-                dateOutStr = f" bis {self.dsc_conf()['state_end'].strftime('%d.%m.%Y')}"
+    @commands.command(name="geck", help="GECKARBOR!")
+    async def geck(self, ctx):
+        await ctx.trigger_typing()
+        file = discord.File(f"{Config().storage_dir(self)}/treeckos.png")
+        await ctx.send("arbor!", file=file)
 
-            embed = discord.Embed(title=f":clipboard: Anmeldung offen{dateOutStr}!")
-            embed.add_field(name="Aktueller Ausrichter", value=hostNick)
-            embed.add_field(name="Anmeldung", value=self.dsc_conf()['contestdoc_link'])
-            await ctx.send(embed=embed)
-
-        elif self.dsc_conf()['state'] == DscState.Voting:
-            if self.dsc_conf()['state_end'] > datetime.now():
-                dateOutStr = f" bis {self.dsc_conf()['state_end'].strftime('%d.%m.%Y, %H:%M')} Uhr"
-
-            embed = discord.Embed(title=f":incoming_envelope: Votingphase läuft{dateOutStr}!")
-            embed.add_field(name="Votings an", value=hostNick)
-            embed.add_field(name="Alle Songs", value=self.dsc_conf()['contestdoc_link'])
-            embed.add_field(name="Youtube-Playlist", value=self.dsc_conf()['yt_link'])
-            await ctx.send(embed=embed)
-
-        else:
-            await ctx.send("Configuration error. Please reset dsc configuration.")
-            embed = discord.Embed(title="DSC configuration error")
-            embed.add_field(name="Host ID", value=str(self.dsc_conf()['host_id']))
-            embed.add_field(name="Host Name", value=hostNick)
-            embed.add_field(name="State", value=str(self.dsc_conf()['state']))
-            embed.add_field(name="YT Playlist", value=str(self.dsc_conf()['yt_link']))
-            embed.add_field(name="State End", value=str(self.dsc_conf()['state_end']))
-            await utils.write_debug_channel(self.bot, embed)
-
-    @dsc.group(name="set", help="Set data about current/next DSC.", usage="<host|state|stateend|yt>")
-    @commands.has_any_role(Config().ADMIN_ROLE_ID, Config().BOTMASTER_ROLE_ID, songmaster_role_id)
-    async def dsc_set(self, ctx):
-        """Basic set subcommand, does nothing"""
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help(self.dsc_set)
-
-    @dsc_set.command(name="host", help="Sets the current/next DSC hoster", usage="<user>")
-    #@commands.has_any_role("mod", "songmaster", "botmaster")
-    async def dsc_set_host(self, ctx, user: discord.Member):
-        """Sets the current/next DSC host"""
-        self.dsc_conf()['host_id'] = user.id
-        Config().save(self)
-        await ctx.send("New hoster set.")
-
-    @dsc_set.command(name="state", help="Sets the current DSC state (Voting/Registration)",
-                     usage="<voting|registration>")
-    #@commands.has_any_role("mod", "songmaster", "botmaster")
-    async def dsc_set_state(self, ctx, state):
-        """Sets the current DSC state (registration/voting)"""
-        if state.lower() == "voting":
-            self.dsc_conf()['state'] = DscState.Voting
-            await ctx.send("Voting phase set.")
-        elif state.lower() == "registration":
-            self.dsc_conf()['state'] = DscState.Registration
-            await ctx.send("Registration phase set.")
-        else:
-            await ctx.send("Invalid DSC phase.")
-        Config().save(self)
-
-    @dsc_set.command(name="yt", help="Sets the Youtube playlist link", usage="<link>")
-    #@commands.has_any_role("mod", "songmaster", "botmaster")
-    async def dsc_set_yt_link(self, ctx, link):
-        """Sets the youtube playlist link"""
-        link = utils.clear_link(link)
-        self.dsc_conf()['yt_link'] = link
-        Config().save(self)
-        await ctx.send("New Youtube playlist link set.")
-
-    @dsc_set.command(name="stateend", help="Sets the registration/voting end date", usage="DD.MM.JJJJ[ HH:MM]",
-                     description="Sets the end date and time for registration and voting phase. "
-                                 "If no time is given, 23:59 will be used.")
-    #@commands.has_any_role("mod", "songmaster", "botmaster")
-    async def dsc_set_state_end(self, ctx, dateStr, timeStr=None):
-        """Sets the end date (and time) of the current DSC state"""
-        if not timeStr:
-            dateStr += " 23:59"
-        self.dsc_conf()['state_end'] = datetime.strptime(dateStr,"%d.%m.%Y %H:%M")
-        Config().save(self)
-        await ctx.send("New state end date set.")
+    @commands.command(name="liebe", help="Provides love to the channel")
+    async def liebe(self, ctx):
+        await ctx.send("https://www.youtube.com/watch?v=TfmJPDmaQdg")
